Moved the markdown logic into a MarkdownEditor class.
